Batch 5

+++ src/main/java/personal/rathercynicalbadger/BoredGameChap/service/BGGAPIService.java
@@ -7,6 +7,6 @@ import java.util.Map;
 public interface BGGAPIService {
     Map<Integer, String> searchByTitle(String title);
 
-    Game fetchGame(Integer bggId);
+    Game fetchGame(Long bggId);
 
 }
